Extract redirect components
